Milestones/CS-340-Pet-Shelter-WebApp/requirements.txt
dash==2.17.0
dash-leaflet==1.0.13
pandas==2.2.2
plotly==5.21.0
matplotlib==3.8.4
numpy==1.26.4
pymongo==4.7.2
mysql-connector-python==8.0.32
python-dotenv==0.21.1